Site is fully functional
The site is now fully functional and works just the same as the project i made in vanilla javascript but now in react! This was a great learning experience and i feel i have a much better handle on react now. I might add some more features to this site later on down the line. Things like adding more search and sort algorithms or maybe add some more visualization to the sorting algorithms so you can see them sorting the array.